updated job finder agent
